# docs/openapi-snapshot.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Kiln API",
    "description" : "基于 Spring Boot 4 + Spring Modulith 的企业级后端脚手架",
    "contact" : {
      "name" : "Kiln Team",
      "email" : "dev@skyflux.com"
    },
    "license" : {
      "name" : "Apache-2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version" : "0.0.1-SNAPSHOT"
  },
  "paths" : {
...
    "/api/v1/orders" : {
      "get" : {
        "tags" : [ "order-controller" ],
        "operationId" : "list_2",
        "parameters" : [ {
          "name" : "query",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/PageQuery"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RPageResultOrderResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "order-controller" ],
        "operationId" : "create_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateOrderRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ROrderResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{id}/ship" : {
      "post" : {
        "tags" : [ "order-controller" ],
        "operationId" : "ship",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ROrderResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/orders/{id}/confirm" : {
      "post" : {
        "tags" : [ "order-controller" ],
        "operationId" : "confirm",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ROrderResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{id}/cancel" : {
      "post" : {
        "tags" : [ "order-controller" ],
        "operationId" : "cancel",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ROrderResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/orders/{id}" : {
      "get" : {
        "tags" : [ "order-controller" ],
        "operationId" : "get_3",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ROrderResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CreateOrderRequest" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OrderItemRequest"
            },
            "minItems" : 1
          },
          "note" : {
            "type" : "string"
          }
        }
      },
      "OrderItemRequest" : {
        "type" : "object",
        "properties" : {
          "productId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "productCode" : {
            "type" : "string",
            "minLength" : 1
          },
          "productName" : {
            "type" : "string",
            "minLength" : 1
          },
          "unitPrice" : {
            "type" : "number",
            "minimum" : 0
          },
          "currency" : {
            "type" : "string",
            "maxLength" : 3,
            "minLength" : 0
          },
          "quantity" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 1
          }
        },
        "required" : [ "productId", "unitPrice" ]
      },
      "OrderItemResponse" : {
        "type" : "object",
        "properties" : {
          "productId" : {
            "type" : "string"
          },
          "productCode" : {
            "type" : "string"
          },
          "productName" : {
            "type" : "string"
          },
          "unitPrice" : {
            "type" : "number"
          },
          "quantity" : {
            "type" : "integer",
            "format" : "int32"
          },
          "subtotal" : {
            "type" : "number"
          }
        }
      },
      "OrderResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OrderItemResponse"
            }
          },
          "totalAmount" : {
            "type" : "number"
          },
          "currency" : {
            "type" : "string"
          },
          "note" : {
            "type" : "string"
          }
        }
      },
      "ROrderResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/OrderResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "PageQuery" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 1
          },
          "size" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 200,
            "minimum" : 1
          },
          "sort" : {
            "type" : "string"
          }
        }
      },
...
      "PageResultOrderResponse" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OrderResponse"
            }
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "RPageResultOrderResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PageResultOrderResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "name" : "Authorization",
        "in" : "header",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}